FIX UI in inconsistent state on nCards change
This was due to ListItem memo. ListItems were not rerendered
when nCards varied because visible content was unaffected.

Now pass a view key along to force rerender when path varies (and
callbacks should be updated).

Also now use nCards, persisted in localStorage instead of
rightOpened. This allows to better recover UI on reload, and
prepares for more than 2 cards.